コンポーネントの読み込みまで完了
