+++ app/views/shared/_articlebox_react_test.html.erb
@@ -21,6 +21,6 @@
   </div>
   <div class="article-preview">
     <h1 class="preview-label">記事プレビュー</h1>
-    <%= javascript_pack_tag 'preview' %>
+    <%=  react_component("preview") %>
   </div>
 </div>
